rename hook file
